# Speed up test by setting gem_details to false