@kbaum 2019-02-25
I think the extra time probably just comes from viewing the report with all the detail?